Include user info in getAll return from TRPCRouter
Use clerkClient, to check for user data on the on posts router.
Map the author  to the post by finding user id matches in post.authorId.
Throw an error, if it doesn't find an author for a post.
Return post data and author data in the same object.
Create a PostView component to render returned data from the api.
Create a PostWizard component with no functionality.
Wizard will be used, so a user can create new posts.